move milvus util
